fixed error of displaying signup result.
We were trying to assign a variable check_error to the
result of the hook signup(), but signup() doesn't return
anything! so when we try to do destructuring assignment
it throws an error. Now, we correctly use the state
error declared in the useSignuo hook to display any
error messages instead of looking for it from signup().